# apps/api/src/todos/dto/create-todo.dto.ts
import { TodoVisibility } from '@prisma/client';
import {
  IsEnum,
  IsOptional,
  IsString,
  MaxLength,
  MinLength,
} from 'class-validator';
import { Trim } from '../../common/decorators/trim.decorator';

export class CreateTodoDto {
  @Trim()
  @IsString()
  @MinLength(3)
  @MaxLength(100)
  title!: string;

  @Trim()
  @IsOptional()
  @IsString()
  @MinLength(3)
  @MaxLength(256)
  description?: string;

  @IsOptional()
  @IsString()
  @IsEnum(TodoVisibility)
  visibility?: TodoVisibility;
}